+++ hdt-lib/include/Triples.hpp
@@ -84,7 +84,7 @@ public:
 	 *
 	 * @return
 	 */
-	virtual unsigned int getNumberOfElements()=0;
+    virtual size_t getNumberOfElements()=0;
 
 	/**
 	 * Returns size in bytes of the overall structure.
