+++ hdt-lib/src/triples/PlainTriples.hpp
@@ -77,7 +77,7 @@ public:
 	 *
 	 * @return
 	 */
-	unsigned int getNumberOfElements();
+    size_t getNumberOfElements();
 
     size_t size();
