+++ hdt-lib/src/triples/TripleListDisk.hpp
@@ -88,7 +88,7 @@ public:
 	 *
 	 * @return
 	 */
-	unsigned int getNumberOfElements();
+    size_t getNumberOfElements();
 
 	/**
 	 * Returns size in bytes
